feat: 7단계 구현
존재하지 않는 명언삭제에 대한 예외처리 기능 추가

+++ src/main/java/com/ll/service/QuotationService.java
@@ -39,8 +39,10 @@ public class QuotationService {
     public void deleteQuotation(Rq rq) {
         int id = rq.parseInt("id", 0);
 
-        quotationRepository.deleteById(id);
-
-        System.out.println(id);
+        if (quotationRepository.existsById(id)) {
+            quotationRepository.deleteById(id);
+        } else {
+            System.out.println(id + "번 명언은 존재하지 않습니다.");
+        }
     }
 }
